+++ frontend/src/utils/components/tasks/WeTaskHeaderActions.tsx
@@ -4,7 +4,6 @@
 import React, { useEffect, useState } from 'react';
 import { BusyIndicator, BusyIndicatorSize, Button, ButtonDesign } from '@ui5/webcomponents-react';
 import '@ui5/webcomponents-icons/dist/synchronize';
-import '@ui5/webcomponents-icons/dist/begin';
 import '@ui5/webcomponents-icons/dist/cancel';
 import '@ui5/webcomponents-icons/dist/user-edit';
 import { useDispatch, useSelector } from 'react-redux';
@@ -92,8 +91,6 @@ const WeTaskHeaderActions: React.FC<AdminTaskHeaderActionsProps> = props => {
     dispatch(postRequest(WeDataKey.ADMIN_UNASSIGN_TASK, { task_id: props.taskId }));
   };
 
-  const handleSkipTask = (): void => {};
-
   const handleTryAgain = (): void => {
     dispatch(postRequest(WeDataKey.ADMIN_EXECUTE_ERRONEOUS_TASK, { task_id: props.taskId }));
   };
@@ -163,18 +160,6 @@ const WeTaskHeaderActions: React.FC<AdminTaskHeaderActionsProps> = props => {
             setUserDialogOpen(true);
           }}
         />
-
-        <BusyIndicator active={false} delay={0} size={BusyIndicatorSize.Small}>
-          <Button
-            icon="begin"
-            design={ButtonDesign.Transparent}
-            tooltip={t('admin.skipTasksTooltip')}
-            disabled={isReadonly}
-            onClick={() => {
-              handleSkipTask();
-            }}
-          />
-        </BusyIndicator>
 
         <BusyIndicator
           active={executeErroneousTaskLoadState}